Changed the layout to split into two columns first then add content.
styles.css: Removed unused classes and tweaked the dimentions of the
columns.
index.html: Rearanged the column dividers and moved the main_body
dividers to only encapsulate the text of the site (not links or footer).
unit_trig.js: Made sketch take up the entire column rather than a
fraction of it.