Added Arcs & Angles page.
Found something really neat: the angle associated with an arc is equal
to the length of the arc divided by its radius. That will probably be a
useful way of framing those values in order to get to a "angle".
Going to add some numbers to each arc and radius to show that the angle
really does stay the same even if the radius and arc length change.