Added triangles page.
The JS currently has a few floating triangles as the animation. The goal
is to have a bunch of smaller triangles orbiting around the screen. With
all different types (obtuse, right, scalene, ect).
Might also want to mark the points of the triangles with dots.
Should also try to put the centers of orbit in different places. Maybe
orbiting things themselves.