Added right-triangles page.
The JS is interesting to look at currently: an arc with radius, a
connecting chord from the start and end of the arc, a triangle formed
from the radius and sine and cosine lines.
Not entirely sure what to use for the JS here. It would be nice to
include non-right triangles, but may also take away from the eventual
goal of talking about trig functions.