Added triangles_and_circles page with mostly complete js.
Decided to use a simple circle with triangle design (the one with
radius as hypotonuse, sine as height, and cosine as width).

Found a consistent way to present text with some notes on a few
parameters that control the strokeWeight, stroke, and fill of the font.
Just using one text function and spacing using newline characters. Using
this to show a table of values for the angle, cos, sin, and the radius.

Also added a graphic of the relitive lengths of each of the values in
the table. Very simply added a line for each value that changes in
length as the value does.