Added directory for trig identities.
Goal is to have visualizations of as many trig identities as is
meaningful.
Starting with just the complex exponential and then building from there.
It feels right to start here because everything else can be built from
it. In a similar way to how other geometry can start from a compass and
straight edge.
Should try to find a clear way to depict adding exponentials or just
representing them graphically.