JS can now toggle between showing exponential or cos/sin components.
Added logic to detect space bar or shift key press. This causes the JS
to switch between showing exp(jtheta) or the sin and cosine components.

Also changed color of origin to be a 128 gray.

Want to find a better colorscheme for the connexting lines. Might use a
grayscale color. That said, contrast between the components and sum
would be nice.