Changed styles.css title to use text-align:center and removed margin of 5em.
It turns out the reason I could not get the title of the pages to center
was because it had been set to have a margin or padding of 5em to the
right. The reason for this was to try and fix an issue that happens when
the page width is made to small where the left and right navigaiton
buttons apear next to eachother instead of around the title.
I had forgotten I'd made this change and did not know why the title
stayed un-centered. Now that this has been fixed the text lines up much
nicer and the issue that it was attempting to fix only really occurs in
unlikely situations.
Eventually it will likely be nessesary to try and come up with a fix
involving detecting the screen size in order to not have the navigaiton
buttons go next to eachother but is low priority.